Added check to see if the message in the queue is empty or not

+++ BackEnd/BackEndQueue.py
@@ -44,17 +44,17 @@ while True:
         MaxNumberOfMessages = 1,
         MessageAttributeNames = ['All']
         )
-    print(cat_response)
-    cat_message = cat_response['Messages'][0]
-    cat = cat_message['Body']
-    receipt_handle = cat_message['ReceiptHandle']
-    sqs.delete_message(
-        QueueUrl = FRONT_TO_BACK_QUEUE,
-        ReceiptHandle = receipt_handle
-    )
-    url = get_URL(s3_client, BUCKET_NAME, cats_list, cat) # Get the URL and send it to FrontEnd.
-    sqs.send_message( # Send the URL to the FrontEnd through the back-to-front SQS queue.
-        QueueUrl = BACK_TO_FRONT_QUEUE,
-        MessageAttributes={},
-        MessageBody= url
-    )
+    if 'Messages' in cat_response:
+        cat_message = cat_response['Messages'][0]
+        cat = cat_message['Body']
+        receipt_handle = cat_message['ReceiptHandle']
+        sqs.delete_message(
+            QueueUrl = FRONT_TO_BACK_QUEUE,
+            ReceiptHandle = receipt_handle
+        )
+        url = get_URL(s3_client, BUCKET_NAME, cats_list, cat) # Get the URL and send it to FrontEnd.
+        sqs.send_message( # Send the URL to the FrontEnd through the back-to-front SQS queue.
+            QueueUrl = BACK_TO_FRONT_QUEUE,
+            MessageAttributes={},
+            MessageBody= url
+        )
